refactor: moved add_user_to_cursus() to a separate file.
